# M2-Challenge-Hoelscher-Rachel/swagger.yaml
swagger: '2.0'
info:
  description: |
    This is a practice API for Module 2.  You can find 
    out more about Swagger at 
    [http://swagger.io](http://swagger.io) or on 
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  version: 1.0.0
  title: Module 2 API
  termsOfService: http://swagger.io/terms/
  contact:
    email: rachel.hoelscher@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# host: petstore.swagger.io
# basePath: /v2
...
paths:
...
  /randomMonth/:
    get:
      tags:
      - month
      summary: Find random month
      description: Returns a random month
      operationId: getRandomMonth
      produces:
      - application/json
      - application/xml
      responses:
        200:
          description: successful operation
        400:
          description: Invalid ID supplied
      security:
      - api_key: []
...
securityDefinitions:
   api_key:
    type: apiKey
    name: api_key
    in: header
...
host: virtserver.swaggerhub.com
basePath: /me0wmerz/Module_2_Challenge/1.0.0
schemes:
 - https
 - http